refactor(rtk): reset all rtk slices to 'camelCase' as should be for all files
